# Use `collections.defaultdict` to speed up metrics aggregation